ルーディングと購入テーブルの作成
