+++ app/helpers/buys_helper.rb
@@ -1,0 +1,2 @@
+module BuysHelper
+end
