+++ app/views/items/show.html.erb
@@ -34,7 +34,7 @@
       <%= link_to '削除', item_path{@items.id}, method: :delete, class:'item-destroy' %>
       <% else %>
       <%# 商品が売れていない場合はこちらを表示しましょう %>
-      <%= link_to '購入画面に進む', "#" ,class:"item-red-btn"%>
+      <%= link_to '購入画面に進む',  ,class:"item-red-btn"%>
       <%# //商品が売れていない場合はこちらを表示しましょう %>
 
       <%# // ログインしているユーザと出品しているユーザが同一人物である時、商品の編集と削除を表示にしましょう %>
