+++ spec/requests/buys_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Buys", type: :request do
+
+end
